Added getters and setters for all settings values and renamed some members.
Some members and functions were named in different ways, now the naming scheme is a bit more unified. Also changed some types to size_t where negative values don't make sense.

+++ src/projectM-sdl/pmSDL.hpp
@@ -152,8 +152,8 @@ private:
     SDL_Window *win{ nullptr };
     bool isFullScreen{ false };
     SDL_AudioDeviceID audioInputDevice{ 0 };
-    unsigned int width{ 0 };
-    unsigned int height{ 0 };
+    size_t width{ 0 };
+    size_t height{ 0 };
     GLuint programID{ 0 };
     GLuint m_vbo{ 0 };
     GLuint m_vao{ 0 };
